// src/types/budget.d.ts
export type BudgetResType = {
  sum: number;
  shopping: number;
  food: number;
  traffic: number;
  hospital: number;
  fee: number;
  education: number;
  leisure: number;
  society: number;
  daily: number;
  overseas: number;
};
export type BudgetReqType = {
  shopping: number;
  food: number;
  traffic: number;
  hospital: number;
  fee: number;
  education: number;
  leisure: number;
  society: number;
  daily: number;
  overseas: number;
};